+++ proto/varlogpb/storage_node.go
@@ -1,14 +1,12 @@
 package varlogpb
 
 func (snd *StorageNodeDescriptor) Valid() bool {
-	if snd == nil ||
-		len(snd.Address) == 0 ||
-		len(snd.Storages) == 0 {
+	if snd == nil || len(snd.Address) == 0 || len(snd.Paths) == 0 {
 		return false
 	}
 
-	for _, storage := range snd.Storages {
-		if !storage.valid() {
+	for _, path := range snd.Paths {
+		if len(path) == 0 {
 			return false
 		}
 	}